Clearify code
Be more brief and define all classes of the
interface in a single source file.
