--- frame.py
@@ -1,13 +1,0 @@
-from typing import NamedTuple
-
-from .status import DaliStatus
-
-
-class DaliFrame(NamedTuple):
-    timestamp: float = 0
-    length: int = 0
-    data: int = 0
-    priority: int = 1
-    send_twice: bool = False
-    status: int = DaliStatus.OK
-    message: str = "OK"
